- [+] use go generate and refactor flag handling to flags.go
