+++ cmd/easygen/config.go-
@@ -1,0 +1,22 @@
+// !!! !!!
+// WARNING: Code automatically generated. Editing discouraged.
+// !!! !!!
+
+package easygen
+
+////////////////////////////////////////////////////////////////////////////
+// Constant and data type/structure definitions
+
+const progname = "easygen" // os.Args[0]
+
+// The Options struct defines the structure to hold the commandline values
+type Options struct {
+	HTML         bool   // treat the template file as html instead of text
+	TemplateStr  string // template string (in text)
+	TemplateFile string // .tmpl (comma-separated) template file `name(s)` (default: same as .yaml file)
+	ExtYaml      string // `extension` of yaml file
+	ExtTmpl      string // `extension` of template file
+	StrFrom      string // replace from, the from string used for the replace function
+	StrTo        string // replace to, the to string used for the replace function
+	Debug        int    // debugging `level`
+}
